Added valid version of hw
